refactor(apigateway/common/tenant): move all constants into common/tenant
