+++ src/dashboard/apigateway/apigateway/common/constants.py
@@ -18,7 +18,6 @@
 #
 
 from blue_krill.data_types.enum import EnumField, StructuredEnum
-from django.utils.translation import gettext_lazy as _
 
 CACHE_TIME_5_MINUTES = 5 * 60
 CACHE_TIME_24_HOURS = 24 * 3600
@@ -30,18 +29,6 @@ CACHE_MAXSIZE = 2000
 class LanguageCodeEnum(StructuredEnum):
     EN = EnumField("en", "en")
     ZH_HANS = EnumField("zh-hans", "zh-hans")
-
-
-TENANT_MODE_SINGLE_DEFAULT_TENANT_ID = "default"
-TENANT_MODE_GLOBAL_DEFAULT_TENANT_ID = ""
-TENANT_ID_OPERATION = "system"
-
-
-class TenantModeEnum(StructuredEnum):
-    """租户模式"""
-
-    GLOBAL = EnumField("global", _("全租户"))
-    SINGLE = EnumField("single", _("单租户"))
 
 
 # IP 或 IP 网段正则
